Rename utils/waitTime.tsx to utils/time.tsx
