+++ src/screens/LocationDetailsScreen.tsx
@@ -15,7 +15,7 @@ import { AntDesign } from "@expo/vector-icons";
 import { ThemeContext } from "@contexts/theme";
 import { StyledText } from "@components/StyledText";
 import { SuccessModal } from "@components/SuccessModal";
-import { renderLastUpdated } from "@utils/waitTime";
+import { renderLastUpdated } from "@utils/time";
 import { LocationDetailsScreenProps } from "@navigators/WaitTimeStackNavigator";
 
 export const LocationDetailsScreen = ({
