+++ src/utils/time.tsx
@@ -15,3 +15,19 @@ export function renderLastUpdated(lastUpdated: number): string {
     return string + "now";
   }
 }
+
+/**
+ * Return a string in the format of "X hrs Y min" or "Y min" depending on the
+ * the number of minutes passed.
+ * @param numMinutes Number of in minutes
+ * @returns string in the format of "X hrs Y min" or "Y min"
+ */
+export const toHoursAndMinutes = (numMinutes: number) => {
+  const hours = Math.floor(numMinutes / 60);
+  const minutes = numMinutes % 60;
+
+  if (hours > 0) {
+    return `${hours} hrs ${minutes} min`;
+  }
+  return `${minutes} min`;
+};
